requirements.txt
appdirs==1.4.4
bonobo==0.6.4
certifi==2020.6.20
chardet==3.0.4
colorama==0.3.9
configparser==5.0.0
crayons==0.4.0
dataclasses==0.7
email-notifier==0.0.1
freezegun==1.0.0
fs==2.4.11
graphviz==0.8.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
mondrian==0.8.0
packaging==19.2
patch==1.16
pbr==5.4.5
psutil==5.7.2
psycopg2-binary==2.8.5
pydantic==1.6.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
python-slugify==1.2.6
pytz==2020.1
requests==2.24.0
selenium==3.141.0
six==1.15.0
SQLAlchemy==1.3.19
stevedore==1.32.0
Unidecode==1.1.1
urllib3==1.25.10
webdriver-manager==3.2.2
whistle==1.0.1
xlrd==1.2.0
xml-stream==0.0.2
